moving templates into markdown package in the vcs directory
