+++ server/vcs/markdown/templates/applyWrappedSuccess.tmpl
