+++ server/vcs/markdown/templates/approveAllProjects.tmpl
