+++ server/vcs/markdown/templates/failureWithLog.tmpl
