+++ server/vcs/markdown/templates/planSuccessUnwrapped.tmpl
