+++ server/vcs/markdown/templates/policyCheckSuccessWrapped.tmpl
