+++ server/vcs/markdown/templates/singleProjectVersionSuccess.tmpl
