+++ server/vcs/markdown/templates/unwrappedErrWithLog.tmpl
