+++ server/vcs/markdown/templates/versionWrappedSuccess.tmpl
